update barcodes to use mixin
