+++ InvenTree/plugin/mixins/__init__.py
@@ -4,6 +4,7 @@ Utility class to enable simpler imports
 
 from ..builtin.integration.mixins import AppMixin, SettingsMixin, ScheduleMixin, UrlsMixin, NavigationMixin
 from ..builtin.action.mixins import ActionMixin
+from ..builtin.barcode.mixins import BarcodeMixin
 
 __all__ = [
     'AppMixin',
@@ -12,4 +13,5 @@ __all__ = [
     'SettingsMixin',
     'UrlsMixin',
     'ActionMixin',
+    'BarcodeMixin',
 ]
